Ticket #974   * Check for missing data before performing actions.

+++ local/browser/src/main/java/com/thinkparity/ophelia/browser/platform/action/contact/DeleteOutgoingEMailInvitation.java
@@ -45,9 +45,13 @@ public final class DeleteOutgoingEMailInvitation extends AbstractAction {
 
         final ContactModel contactModel = getContactModel();
         final OutgoingEMailInvitation invitation = contactModel.readOutgoingEMailInvitation(invitationId);
-        if (browser.confirm("ContactOutgoingEMailInvitationDelete.ConfirmDeleteMessage",
-                new Object[] {invitation.getInvitationEMail()})) {
-            contactModel.deleteInvitation(invitation);
+        if (null == invitation) {
+            logger.logInfo("Invitation no longer exists.");
+        } else {
+            if (browser.confirm("ContactOutgoingEMailInvitationDelete.ConfirmDeleteMessage",
+                    new Object[] {invitation.getInvitationEMail()})) {
+                contactModel.deleteInvitation(invitation);
+            }
         }
     }
